Update stock entity

+++ src/main/java/tech/investment/project/dto/AccountStockRetrieve.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -11,5 +13,5 @@ public class AccountStockRetrieve {
 
     private String stockId;
     private Integer quantity;
-    private double value;
+    private BigDecimal value;
 }
